+++ .gitignore
@@ -38,3 +38,5 @@ next-env.d.ts
 
 *.spec.ts
 package-lock.json
+.replit
+replit.nix
