--- genesis-vocali/README.md
@@ -1,29 +1,0 @@
-## Getting Started
-
-First, populate environment variables
-
-```
-// .env
-NEXT_PUBLIC_SUPABASE_URL=<your supabase project url>
-NEXT_PUBLIC_SUPABASE_KEY=<your supabase api key>
-```
-
-Then, run the development server:
-
-```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
-```
-
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
-
-More resources:
-
-- [Architecture](https://whimsical.com/frontend-arquitecture-NfUQur2e1WpvWmeH9d1mXU)
-- [Genesis wiki](https://www.notion.so/tianlu/Genesis-Wiki-319c67a41c5a4b9cb5f9eadd6f04e5bb)
-- [Supabase API docs](https://supabase.com/dashboard/project/kxchqnumsodkzillsjhx/api)
