--- genesis-vocali/jest.config.ts
@@ -1,17 +1,0 @@
-import type { Config } from "jest";
-import nextJest from "next/jest";
-
-const createJestConfig = nextJest({
-  dir: "./",
-});
-
-const config: Config = {
-  coverageProvider: "v8",
-  testEnvironment: "jest-environment-jsdom",
-  modulePaths: ["<rootDir>/"],
-  moduleNameMapper: {
-    "^@/(.*)$": "<rootDir>/$1",
-  },
-};
-
-export default createJestConfig(config);
