--- genesis-vocali/next.config.mjs
@@ -1,9 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  i18n: {
-    locales: ['en', 'es'],
-    defaultLocale: 'en'
-  }
-};
-
-export default nextConfig;
