--- genesis-vocali/pages/_app.tsx
@@ -1,20 +1,0 @@
-import "@/styles/globals.css";
-import { AppProps } from "next/app";
-import { NextIntlClientProvider } from "next-intl";
-import { useRouter } from "next/router";
-
-function App({ Component, pageProps }: AppProps) {
-  const router = useRouter();
-
-  return (
-    <NextIntlClientProvider
-      locale={router.locale}
-      timeZone='Europe/Vienna'
-      messages={pageProps.messages}
-    >
-      <Component {...pageProps} />
-    </NextIntlClientProvider>
-  );
-}
-
-export default App;
