--- genesis-vocali/resources/utils/config.ts
@@ -1,4 +1,0 @@
-export default {
-  SUPABASE_URL: process.env.NEXT_PUBLIC_SUPABASE_URL,
-  SUPABASE_API_KEY: process.env.NEXT_PUBLIC_SUPABASE_KEY,
-};
