+++ modules/lang/props.ts
