+++ next.config.mjs
@@ -1,4 +1,9 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  i18n: {
+    locales: ['en', 'es'],
+    defaultLocale: 'en'
+  }
+};
 
 export default nextConfig;
