+++ package.json
@@ -12,13 +12,14 @@
   "dependencies": {
     "@supabase/supabase-js": "^2.43.5",
     "next": "14.2.4",
+    "next-intl": "^3.15.2",
     "react": "^18",
     "react-dom": "^18"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
-    "@types/node": "^20",
-    "@types/react": "^18",
+    "@types/node": "^20.14.9",
+    "@types/react": "^18.3.3",
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.4",
