+++ pages/_app.tsx
@@ -1,6 +1,20 @@
 import "@/styles/globals.css";
-import type { AppProps } from "next/app";
+import { AppProps } from "next/app";
+import { NextIntlClientProvider } from "next-intl";
+import { useRouter } from "next/router";
 
-export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+function App({ Component, pageProps }: AppProps) {
+  const router = useRouter();
+
+  return (
+    <NextIntlClientProvider
+      locale={router.locale}
+      timeZone="Europe/Vienna"
+      messages={pageProps.messages}
+    >
+      <Component {...pageProps} />
+    </NextIntlClientProvider>
+  );
 }
+
+export default App;
