+++ pages/auth/login.tsx
@@ -1,7 +1,13 @@
 import React from "react";
+import { useTranslations } from "next-intl";
+import { getStaticPropsWithTranslations } from "@/modules/lang/props";
+
+export const getStaticProps = getStaticPropsWithTranslations;
 
 const Login: React.FC = () => {
-  return <h1>Login Page</h1>;
+  const t = useTranslations("common");
+
+  return <h1>{t("login")}</h1>;
 };
 
 export default Login;
